working on the liner.py file
