+++ students/g_rama/lesson07/src/linear.py
@@ -1,0 +1,199 @@
+"""Mongo DB class to import the CSV data and to display the data"""
+import csv
+import os
+import pymongo
+from pymongo import MongoClient
+
+
+class MongoDBConnection:
+    """MongoDB Connection"""
+
+    def __init__(self, host='localhost', port=27017):
+        """ be sure to use the ip address not name for local windows"""
+        self.host = host
+        self.port = port
+        self.connection = None
+
+    def __enter__(self):
+        self.connection = MongoClient(self.host, self.port)
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.connection.close()
+
+
+def import_data(directory_name, product_file, customer_file, rental_file):
+    """Import data for inventory management"""
+
+    # mongo = MongoDBConnection()
+    # with mongo:
+    #     # mongodb database; it all starts here
+    #     DB = mongo.connection.hpnorton
+    #     products_error = 0
+    #     customers_error = 0
+    #     rentals_error = 0
+    #     try:
+    #         product_file_csv = os.path.join(directory_name, product_file)
+    #     except FileNotFoundError:
+    #         products_error += 1
+    #     try:
+    #         customer_file_csv = os.path.join(directory_name, customer_file)
+    #     except FileNotFoundError:
+    #         customers_error += 1
+    #     try:
+    #         rental_file_csv = os.path.join(directory_name, rental_file)
+    #     except FileNotFoundError:
+    #         rentals_error += 1
+    #
+    #     products = DB["products"]
+    #     customers = DB["customers"]
+    #     rentals = DB["rentals"]
+    #
+    #     with mongo:
+    #         DB = mongo.connection.hpnorton
+    #         print("Printing the collection of DB")
+    #         print(DB.list_collection_names())
+    #
+    #     try:
+    #         with open(product_file_csv, encoding='utf-8-sig') as product:
+    #             products_csv = csv.DictReader(product)
+    #             for row in products_csv:
+    #                 print(row)
+    #                 try:
+    #                     products.insert_one(row)
+    #                 except pymongo.errors.DuplicateKeyError:
+    #                     products_error += 1
+    #     except FileNotFoundError:
+    #         products_error += 1
+    #
+    #     try:
+    #         with open(customer_file_csv, encoding='utf-8-sig') as customer:
+    #             customers_csv = csv.DictReader(customer)
+    #             for row in customers_csv:
+    #                 print(row)
+    #                 try:
+    #                     customers.insert_one(row)
+    #                 except pymongo.errors.DuplicateKeyError:
+    #                     customers_error += 1
+    #     except FileNotFoundError:
+    #         customers_error += 1
+    #
+    #     try:
+    #         with open(rental_file_csv, encoding='utf-8-sig') as rental:
+    #             rentals_csv = csv.DictReader(rental)
+    #             for row in rentals_csv:
+    #                 print(row)
+    #                 try:
+    #                     rentals.insert_one(row)
+    #                 except Exception as ex:
+    #                     rentals_error += 1
+    #     except FileNotFoundError:
+    #         rentals_error += 1
+    #
+    #     products_count = DB.products.count_documents({})
+    #     customers_count = DB.customers.count_documents({})
+    #     rentals_count = DB.rentals.count_documents({})
+
+    products_count, products_error = import_generic(directory_name, product_file, "products")
+    customers_count, customers_error = import_generic(directory_name, customer_file, "customers")
+    rentals_count, rentals_error = import_generic(directory_name, rental_file, "rentals")
+
+    tuple1 = (products_count, customers_count, rentals_count)
+    tuple2 = (products_error, customers_error, rentals_error)
+    return tuple1, tuple2
+
+
+def import_generic(directory_name, import_file, imported_table):
+    mongo = MongoDBConnection()
+    with mongo:
+        # mongodb database; it all starts here
+        DB = mongo.connection.hpnorton
+        imported_error = 0
+        try:
+            import_file_csv = os.path.join(directory_name, import_file)
+            print(import_file_csv)
+        except FileNotFoundError:
+            imported_error += 1
+
+        imported_table = DB[imported_table]
+        print(DB.list_collection_names())
+
+        # with mongo:
+        #     DB = mongo.connection.hpnorton
+        #     print("Printing the collection of DB")
+        #     print(DB.list_collection_names())
+
+        try:
+            with open(import_file_csv, encoding='utf-8-sig'):
+                imported_table_csv = csv.DictReader(import_file_csv)
+                for row in imported_table_csv:
+                    print(row)
+                    try:
+                        imported_table.insert_one(row)
+                    except pymongo.errors.DuplicateKeyError:
+                        imported_error += 1
+        except FileNotFoundError:
+            imported_error += 1
+
+        imported_table_count = DB.imported_table.count_documents({})
+    return imported_error, imported_table_count
+
+
+def import_customers():
+    pass
+
+
+def import_products():
+    pass
+
+
+def import_rentals():
+    pass
+
+
+def show_available_products():
+    """Display the products in inventory"""
+    mongo = MongoDBConnection()
+    with mongo:
+        DB = mongo.connection.hpnorton
+        print(DB.list_collection_names())
+        available_products = {}
+        for prod in DB.products.find():
+            if int(prod["quantity_available"]) > 0:
+                available_products.update({prod["product_id"]: {prod["description"],
+                                                                prod["product_type"],
+                                                                prod["quantity_available"]}})
+
+        return available_products
+
+
+def show_rentals(product_id):
+    """Display the users who rented a product"""
+    mongo = MongoDBConnection()
+    with mongo:
+        DB = mongo.connection.hpnorton
+        rented_user_id = []
+        rentals_all = DB.rentals.find({'product_id': {'$eq': product_id}})
+        for rental in rentals_all:
+            rented_user_id.append(rental['user_id'])
+        print(rented_user_id)
+        rented_user_info = {}
+        for user in rented_user_id:
+            for rented_user in DB.customers.find({'user_id': {'$eq': user}}):
+                rented_user_info.update({rented_user["user_id"]: {rented_user["name"],
+                                                                  rented_user["address"],
+                                                                  rented_user["email"]}})
+        return rented_user_info
+
+
+def drop_collections():
+    """Function to clean the collections created"""
+    mongo = MongoDBConnection()
+    with mongo:
+        DB = mongo.connection.hpnorton
+        DB.products.drop()
+        print("deleted the products")
+        DB.customers.drop()
+        print("deleted the customers")
+        DB.rentals.drop()
+        print("deleted the rentals")
